Add weight edge tracking to logflow
This change tracks how many times an edge is added and uses this
count as the weight of the edge line.

In my testing setting a large weight isn't making the line any
thicker, so added a count to the label as well. E.G:

  PUT (201) [1]

Where the '1' is the number of times this edge was added,
therefore called in the log.

The python version doesn't seem to allow me to get the edge by
nodes and label, so can't modify the edge label with the count.
So now the edges are collected and then written out after the file
is parsed.